+++ imars_etl/id_lookup.py
@@ -57,7 +57,11 @@ def _id_lookup(
             value
         )
     )
-    return translation[0]
+    # expect only one result
+    assert len(translation) == 1
+    assert len(translation[0]) == 1
+
+    return translation[0][0]
 
 
 def _test_id_lookup(value=None, table=None, first=False):